third_party/blink/web_tests/wpt_internal/bluetooth/descriptor/writeValue/gen-io-op-device-disconnects-before.https.html

<!-- Generated by //third_party/blink/web_tests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-test.js"></script>
<script src="/bluetooth/resources/bluetooth-fake-devices.js"></script>
<script src="/common/gc.js"></script>
<script>
'use strict';
function createDOMException(func) {
  return new DOMException(
      `Failed to execute '${func}' on 'BluetoothRemoteGATTDescriptor': ` +
      `GATT Server is disconnected. Cannot perform GATT operations. ` +
      `(Re)connect first with \`device.gatt.connect\`.`,
      'NetworkError');
}

bluetooth_test(() => {
  let val = new Uint8Array([1]);
  return setBluetoothFakeAdapter('DisconnectingHealthThermometerAdapter')
      .then(() => requestDeviceWithTrustedClick({
              filters: [{services: ['health_thermometer']}],
              optionalServices: [request_disconnection_service_uuid]
            }))
      .then(device => device.gatt.connect())
      .then(gattServer => {
        let measurement_interval;
        return gattServer.getPrimaryService('health_thermometer')
            .then(ht => ht.getCharacteristic('measurement_interval'))
            .then(
                characteristic =>
                    characteristic.getDescriptor(user_description.name))
            .then(d => user_description = d)
            .then(() => get_request_disconnection(gattServer))
            .then(requestDisconnection => requestDisconnection())
            .then(
                () => assert_promise_rejects_with_message(
                    user_description.writeValue(val),
                    createDOMException('writeValue')));
      });
}, 'Device disconnects before writeValue. Reject with NetworkError.');

</script>
